Dev benchmark (#480)
* Make hetnets great again

* Better logging for validation performance

* Add overrides for e.g. Bayesian

* Add deup method including training function

* Remove ty warnings

* ImageNet fix

* Tighter prior for pretrained

* Add credal bnn representer

* Add credal bnn training

* Minor change for credal_net

* Fix SNGP

* Add maximin decision rule

* Add sngp training

* Improve sngp

* Use decide in selective prediction

* Add deup implementation

* Fix laplace

* Change evidential to use clipped exp activatio n

* Add warning for laplace change

* Override edl kl weight for imagenet

* Fix for sngp

* Add option to keep pretrained weights to credal net

* Add ood datasets that are torchvision native

* Improve efficiency of deup fitting

* Add additional ood datasets

* Add key for measure, fix saving in ood

* Fix subensemble registration

* Add some prints

* Fix data loader issue of subsensemble

* Fix credal net not using pretrained weights for imagenet

* Fix deup cpu issue

* Fix for ninco ssb hard

* Improve deup efficiency for imagenet

* Update dropout label

* Fix deup training

* Fix data ood

* Fix deup training issue

* Fix credal net accuracy for imagenet

* Reduce batch_size for laplace fitting

* Fix for ninco, ssb_hard

* Add natural posterior network uncertainty measure vacuity

* Fix for subensemlbe

* Add uncertainty for posterior networks

* Add uncertainty for evidential classification

* Add sngp uncertainty decomposition

* Fix for laplace

* Fix uncertainty decompostions

* Helper function

* Add first order data collection for benchmarking

* Fix deup

* Uncertainty decomp for evidential classification, posterior network, natural posterior network

* Add dare uncertainty

* Revert adding predict_raw

* Fix batching stripping markers

* Minor fix decisions

* Add dare decomposition

* Add TORCH_SAMPLE back in

* Add TORCH_SAMPLE back in

* Increase timeout to 60

* Add select_uncertainty to AL

* Add first order data experiment

* cleanups

* fixes

* Laplace fix

* Speedup with amp

* dispatch workaround
a proper of dare would require the introduction of a proper LogitDistribution type

* fix stale __all__

* Fix device

* added SNGPRepresentation for automatic decompose dispatch

* Fix bayesian and duq downstream issue

* fixed embedding registration

* Fix subensemble decide

* Fix batchensemble representer registration

* Use coverage and efficiency metrics for first order

* split ArrayCategoricalDistribution into probability- and logit-based representations

* split TorchCategoricalDistribution representation
aligns with the array categorical split

* fixed dare

* Fix laplace selective prediction issues and modify run naming for credal alpha methods

* Remove duplicate unnormalized_probabilities property

* Fix laplace test

* Fix convex hull coverage test

* Maybe subensemble fix

* Fix for subensemble

* Update config for downstream tasks

* Update laplace imagenet config

* Update downstream task computation and plotting

* Add batchensemble representer fix

* Fix for batchensemble, should make it faster

* Fix laplace maybe

* Implement reuse of base model for effcre

* Make laplace faster

* Add training for conformalized credal predictors

* Align loading with others

* Improve coverage and efficiency

* Fix for conformalized credal

* Fix rounding

* Fix first order evaluation

* Possibly better hyperparams for natpn and evidential

* Remove print statements

* Fix interval probability calculation to clip negative remaining probabilities

* Fix conformalized credal set predictor calibration

* Allow evaluation of conformal credal set

* Allow evaluation of conformal credal set

* Add subsampling for Laplace

* Fix model loading to handle cal_split_suffix

* Fix loading to device for laplace

* Correct efficiency for conformalized credal sets (with TVD)

* Fix devic

* Add wandb saving fix for Laplace

* Fix loading for laplace

* Safeguard for baselaplace and amp

* Fix subensemble

* Fix logging of id_scores for alpha methods

* Add hydra multirun folder to gitignore

* Fix evidential config

* FIx calibration loading

* Align epsilon value in convex_hull_coverage to interval coverage rounding behaviour

* Fix dare

* Update dare config

* Update dare config

* Fix CreRL

* Add computation of both coverage metrics

* Fix ruff

---------

+++ src/probly/quantification/measure/sample/torch.py
@@ -17,7 +17,7 @@ def torch_mean_squared_distance_to_scaled_one_hot(
     sample: TorchCategoricalDistributionSample, scale: float | None = None
 ) -> torch.Tensor:
     r"""Torch impl. Uses :math:`\|h_k - s e_c\|^2 = \|h_k\|^2 - 2s \max_j h_{k,j} + s^2` (no one-hot built)."""
-    tensor = sample.tensor.logits
+    tensor = sample.tensor.logits.float()
     num_classes = tensor.shape[-1]
     target_scale = float(num_classes) if scale is None else float(scale)
 
@@ -33,5 +33,5 @@ def torch_mean_squared_distance_to_scaled_one_hot(
 @total_logit_sample_variance.register(TorchCategoricalDistributionSample)
 def torch_total_logit_sample_variance(sample: TorchCategoricalDistributionSample) -> torch.Tensor:
     """Torch impl. Variance of total logits (logits summed across members)."""
-    tensor = sample.tensor.logits
+    tensor = sample.tensor.logits.float()
     return torch.var(tensor, dim=sample.sample_dim).sum(dim=-1)
